dom/src/threads/test/test_recursion.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads
-->
<head>
  <title>Test for DOM Worker Threads Recursion</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  const testCount = 2;

  var worker = new Worker("recursion_worker.js");

  var count = 0;
  worker.onerror = function(event) {
    is(event.message, "too much recursion");
    if (++count == testCount) {
      SimpleTest.finish();
    }
  }

  for (var i = 0; i < testCount; i++) {
    worker.postMessage("start");
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
